+++ packages/talker/example/talker_example.dart
@@ -16,19 +16,19 @@ Future<void> main() async {
     logLevel: LogLevel.critical,
   );
 
-  talker.fine('Log fine');
   talker.error('Log error');
   talker.good('Log good');
   talker.verbose('Log verbose');
   talker.warning('Log warning');
   talker.critical('Log critical');
 
-  final httpLog = HttpTalkerLog('Http good');
+  final httpLog = HttpExampleTalkerLog('Http good');
   talker.logTyped(httpLog);
+  talker.logTyped(AnalyticsExampleTalkerLog(talker));
 }
 
-class HttpTalkerLog extends TalkerLog {
-  HttpTalkerLog(String message) : super(message);
+class HttpExampleTalkerLog extends TalkerLog {
+  HttpExampleTalkerLog(String message) : super(message);
 
   @override
   AnsiPen get pen => AnsiPen()..xterm(49);
@@ -36,5 +36,26 @@ class HttpTalkerLog extends TalkerLog {
   @override
   String generateTextMessage() {
     return pen.write(message);
+  }
+}
+
+class AnalyticsExampleTalkerLog extends TalkerLog {
+  AnalyticsExampleTalkerLog(this.talker) : super('Report');
+
+  final Talker talker;
+
+  @override
+  String get title => 'ANALYTICS';
+
+  @override
+  AnsiPen get pen => AnsiPen()..xterm(121);
+
+  @override
+  String generateTextMessage() {
+    final msg = '''
+$displayTitleWithTime$message
+Registred logs: ${talker.history.length}
+Registred symbols: ${talker.history.join().split('').length}''';
+    return msg;
   }
 }