+++ packages/talker_dio_logger/lib/talker_dio_logger_interceptor.dart
@@ -15,10 +15,6 @@ class TalkerDioLogger extends Interceptor {
     this.addonId,
   }) {
     _talker = talker ?? Talker();
-    _talker.registerAddon(
-      code: addonId ?? TalkerOriginalAddons.talkerDioLogger.code,
-      addon: this,
-    );
   }
 
   late Talker _talker;