+++ packages/talker_flutter/lib/src/extensions/history.dart
@@ -6,9 +6,7 @@ extension HistoryListFlutterText on List<TalkerDataInterface> {
   String get flutterText {
     final sb = StringBuffer();
     for (final data in this) {
-      final text = data is FlutterTalkerDataInterface
-          ? data.generateFlutterTextMessage()
-          : data.generateTextMessage();
+      final text = data.generateTextMessage();
       sb.write('$text\n');
       sb.write(ConsoleUtils.getUnderline(30));
     }