+++ packages/talker_flutter/lib/src/models/abstract/abstract.dart
@@ -1,1 +1,0 @@
-export 'talker_flutter_adapter_interface.dart';