+++ packages/talker_flutter/lib/src/ui/widgets/filter/filter.dart
@@ -1,1 +1,0 @@
-export 'talker_screen_filter.dart';