+++ packages/talker_logger/lib/src/models/models.dart
@@ -1,2 +1,2 @@
-export 'log_details/log_details.dart';
-export 'log_level/log_level.dart';
+export 'log_details.dart';
+export 'log_level.dart';