+++ packages/talker_dio_logger/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0-dev.1
+- Upgrade talker to 3.0.0-dev.4 version
+- Make titles well known 
+  (WellKnownTitles.httpRequest.title, WellKnownTitles.httpResponse.title, WellKnownTitles.httpError.title)
+
 # 1.3.0
 - **FEAT**: Update talker_dio_logger dio version to 5.0.0