+++ packages/talker_logger/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 3.0.0-dev.3
+- Move all important files on upper level and simplify folders navigation
+
+# 3.0.0-dev.2
+
+- Make LogLevel.fine deprecatd
+
+# 3.0.0-dev.1
+
+- Remove **TalkerLoggerInterface**
+- Make fine log method Deprecated
+- Rename LogLevelTalkerLoggerFilter -> LogLevelFilter
+- Rename TalkerLoggerFilter -> LoggerFilter
+
 # 2.2.1
 - **FIX**: Fix color reset in console